backend/pyproject.toml
[project]
name = "graspmind"
version = "0.1.0"
description = "AI-Powered Study Platform — Backend API"
readme = "README.md"
authors = [
    { name = "Abhijeet Singh", email = "masterabhijeetsingh@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    # ── Web Framework ────────────────────────────────────────
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # ── Auth & Security ──────────────────────────────────────
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.18",
    # ── Database ─────────────────────────────────────────────
    "supabase>=2.13.0",
    "sqlalchemy>=2.0.40",
    # ── Redis ────────────────────────────────────────────────
    "redis[hiredis]>=6.0.0",
    # ── AI & RAG ─────────────────────────────────────────────
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-qdrant>=0.4.0",
    "llama-index-embeddings-google>=0.3.0",
    "qdrant-client>=1.13.0",
    "google-genai>=1.0.0",
    # ── Document Parsing ─────────────────────────────────────
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    # ── Task Queue ───────────────────────────────────────────
    "taskiq[redis]>=0.11.0",
    "taskiq-redis>=1.0.0",
    # ── Utilities ────────────────────────────────────────────
    "pydantic-settings>=2.8.0",
    "httpx>=0.28.0",
    "bleach>=6.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.0",
    "ruff>=0.11.0",
    "bandit>=1.8.0",
    "pip-audit>=2.9.0",
]

[build-system]
requires = ["uv_build>=0.11.9,<0.12.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "S",    # bandit (security)
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "T20",  # flake8-print
]
ignore = ["S101", "B008"]  # allow assert in tests, ignore Depends() issue

[tool.ruff.lint.isort]
known-first-party = ["graspmind"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "bandit>=1.9.4",
    "pip-audit>=2.10.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.12",
]
